ci: fix failures on Dependabot PRs
Dependabot PRs are treated as forks and don't have access to repo
secrets, including the secret required to configure AWS access and
perform the lexicon validation. Since the validate job does not do
anything that *doesn't* require AWS access, the simplest option is to
skip it entirely if Dependabot is the actor.

This does mean that when Dependabot opens a PR to update either
`actions/checkout` or `aws-actions/configure-aws-credentials`, breaking
changes in these actions will not be surfaced until the PR is merged.
However this is also the status quo, since without the needed
permissions the job fails immediately. The only real difference is
whether GitHub treats CI as having passed or failed — in both cases,
nothing is being (or can be) tested.

A logical improvement here would be adding "local" XML format/schema
validation as a CI job, which would not require AWS permissions. This
is left as an exercise for the reader.